Part2/requirements.txt
streamlit==1.33.0
pdfplumber==0.10.3
pinecone-client==3.2.2
pytesseract==0.3.10
pdf2image==1.16.3
sentence-transformers==2.5.1
langchain==0.1.16
openai>=1.12.0
pillow==10.2.0
python-dotenv==1.0.0
streamlit==1.33.0
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2  # Add this line